+++ runner/anvil_runner/__init__.py
@@ -1,3 +1,3 @@
 """Anvil privileged runner."""
 
-__version__ = "1.4.3"
+__version__ = "1.4.4"